finish category

+++ app/Http/Controllers/Category/CategoryController.php
@@ -7,11 +7,17 @@ use App\Models\Category;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use App\Policies\CategoryPolicy;
 use App\Http\Requests\Category\CreateRequest;
 use App\Http\Requests\Category\UpdateRequest;
 
 class CategoryController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->authorizeResource(Category::class, 'category', ['except' => ['index', 'show']]);
+    }
 
     public function index(Request $request)
     {
@@ -80,6 +86,8 @@ class CategoryController extends Controller
 
     public function restore($id)
     {
+        $this->authorize('restore', Category::class);
+
         $category = Category::onlyTrashed()
             ->where('id', $id)
             ->restore();
@@ -96,8 +104,10 @@ class CategoryController extends Controller
         ]);
     }
 
-    public function delete($id)
+    public function forceDelete($id)
     {
+        $this->authorize('forceDelete', Category::class);
+
         $category = Category::onlyTrashed()
             ->where('id', $id)
             ->forceDelete();
